+++ src/dynamics/drag.jl
@@ -88,7 +88,7 @@ function drag!(
     model::AbstractModel,
 )
     (; vor_tend) = diagn.tendencies
-    vor = progn.vor[1]
+    vor = get_step(progn.vor, lf)
 
     # scale by radius (but only once, the second radius is in vor)
     c = drag.c * diagn.scale[]
@@ -145,7 +145,7 @@ function drag!(
     lf::Integer,
     model::AbstractModel,
 )
-    vor = progn.vor[lf]
+    vor = get_step(progn.vor, lf)
     (; vor_tend) = diagn.tendencies
     (; ζ₀) = drag